redirect logs to specific appender
