+++ RelationalAI.Test/log4net.config
@@ -1,13 +1,3 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <log4net>
-  <appender name="RAIAppender" type="RelationalAI.Test.TestOutputAppender">
-    <level value="INFO" />
-    <layout type="log4net.Layout.PatternLayout">
-      <conversionPattern value="%date [%thread] %-5level %logger - %message%newline" />
-   </layout>
-  </appender>
-  <root>
-    <level value="ALL"/>
-    <appender-ref ref="RAIAppender" />
-  </root>
 </log4net>
